// src/app/core/Models/UserInfo.Model.ts
export interface UserInfo {
    userName: String;
    firatName: string;
    lastName: string;
    email: string;
    mobileNumber: string;
    password: string;
}

export interface UserDetails {
    userId: string,
    name: string;
    address: string;
    mobileNumber: string;
    oldPassword: string;
    newPassword: string;
}