// src/app/core/Enums/UrlEnums.ts
export enum UrlEnums {
    MAIN_API = 'http://irentaa.ap-southeast-2.elasticbeanstalk.com/api/',
    LOGIN_CONTROLLER = 'Login',
    HOME_CONTROLLER = 'home',
    PRODUCT_CONTROLLER = 'Products',
    PRODUCT_RATING_CONTROLLER = 'ProductRatings',
    PRODUCT_RENTNOW = 'RentNow',
    SELLER_RENTALS_CONTROLLER = 'SellerRentals',
    SHOPPINGCART_CONTROLLER = 'ShoppingCart',
    SHOPPINGCARTITEMS_CONTROLLER = 'ShoppingCartItems',
    USER_CONTROLLER = 'User',
    TRANSACTION_CONTROLLER = 'transactions',
    PRODUCT_CATAGORIES_CONTROLLER = 'ProductCategories',
    POSTAL_CODES_CONTROLLER = 'PostalCodes',
    BUYER_PURCHASE_CONTROLLER = 'BuyerPurchases',
    CATAGARIES_CONTROLLER = 'category',
    STATUS_CHANGE = 'statuschange/CheckQR',
   
    
}

export enum FileStoreUrlEnums {
    MAIN_BUCKET = 'https://irentaa-dev-content.s3-ap-southeast-2.amazonaws.com/',
    PRODUCT_IMAGES_PATH = 'AppContent/Images/productImages/',
    CATAGORY_IMAGES_PATH = 'AppContent/Images/catagoryImages/',
    PROFILE_IMAGES_PATH = 'AppContent/Images/profilePictures/'
}